fix: server struct
